Handle already exited ogg123(1) gracefully
`player.c` expected the ogg123(1) instance to be running when
`plr_stop` is called. Anticipate that the ogg123 might already
have exited.

 * handle already exited ogg123 gracefully

+++ player.c
@@ -8,6 +8,7 @@
 
 #include <assert.h>
 #include <err.h>
+#include <errno.h>
 #include <signal.h>
 #include <stdlib.h>
 #include <string.h>
@@ -89,7 +90,7 @@ void
 plr_stop(struct plr_context *ctx)
 {
         if (ctx->pid != (pid_t) (-1)) {
-                if (kill(ctx->pid, SIGHUP))
+                if ((kill(ctx->pid, SIGHUP) == -1) && (errno != ESRCH))
                         err(EX_OSERR, "could not kill running player: %d", ctx->pid);
                 ctx->pid = -1;
         }
